+++ tests/.gitignore
@@ -7,3 +7,5 @@ decimal
 define_parse
 define_replace
 enum
+formulas
+
